Update cisbed defaults to use the same as ciscyc.
This results in a tiny bit larger areas with GRIP as with EPICA. But the added
area comes mainly from peripheral glaciers and ice caps, e.g. from the Brooks
Range, while the glaciated areas of the Cordilleran ice sheet itself are
comparable. So I am quite happy with this result, and it will be easy to defend
as it is based on the previous study.